moved the view model to its own file
