+++ todo_app/app.py
@@ -4,28 +4,13 @@ from todo_app.data import session_items
 from todo_app.flask_config import Config
 from flask import request
 import requests
+from todo_app.view_model import ViewModelgit
 
 app = Flask(__name__)
 app.config.from_object(Config)   
 
 KEY=os.getenv("KEY")
 TOKEN=os.getenv("TOKEN")
-
-class ViewModel:
-    def __init__(self, items):
-        self._items = items
-
-    @property
-    def itemsToDo(self):
-        return get_ToDo_Trello(self._items)
-
-    @property
-    def itemsDoing(self):
-        return get_Doing_Trello(self._items)
-
-    @property
-    def itemsDone(self):
-        return get_Done_Trello(self._items)  
 
 @app.route('/index')
 def index():
